requirements.txt
annotated-types==0.7.0 ; python_version >= "3.13"
anyio==4.8.0 ; python_version >= "3.13"
audioop-lts==0.2.1 ; python_version >= "3.13"
blinker==1.9.0 ; python_version >= "3.13"
certifi==2025.1.31 ; python_version >= "3.13"
cffi==1.17.1 ; python_version >= "3.13"
charset-normalizer==3.4.1 ; python_version >= "3.13"
click==8.1.8 ; python_version >= "3.13"
colorama==0.4.6 ; python_version >= "3.13" and platform_system == "Windows"
contourpy==1.3.1 ; python_version >= "3.13"
cycler==0.12.1 ; python_version >= "3.13"
distro==1.9.0 ; python_version >= "3.13"
flask-login==0.6.3 ; python_version >= "3.13"
flask-sqlalchemy==3.1.1 ; python_version >= "3.13"
flask==3.1.0 ; python_version >= "3.13"
fonttools==4.56.0 ; python_version >= "3.13"
greenlet==3.1.1 ; python_version == "3.13" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32")
h11==0.14.0 ; python_version >= "3.13"
httpcore==1.0.7 ; python_version >= "3.13"
httpx==0.28.1 ; python_version >= "3.13"
idna==3.10 ; python_version >= "3.13"
itsdangerous==2.2.0 ; python_version >= "3.13"
jinja2==3.1.6 ; python_version >= "3.13"
jiter==0.9.0 ; python_version >= "3.13"
joblib==1.4.2 ; python_version >= "3.13"
kiwisolver==1.4.8 ; python_version >= "3.13"
lightgbm==4.6.0 ; python_version >= "3.13"
markupsafe==3.0.2 ; python_version >= "3.13"
matplotlib==3.10.1 ; python_version >= "3.13"
mysql-connector-python==9.2.0 ; python_version >= "3.13"
narwhals==1.31.0 ; python_version >= "3.13"
numpy==2.2.3 ; python_version >= "3.13"
openai==1.66.3 ; python_version >= "3.13"
packaging==24.2 ; python_version >= "3.13"
pandas==2.2.3 ; python_version >= "3.13"
patsy==1.0.1 ; python_version >= "3.13"
pillow==11.1.0 ; python_version >= "3.13"
plotly-express==0.4.1 ; python_version >= "3.13"
plotly==6.0.1 ; python_version >= "3.13"
pyaudio==0.2.14 ; python_version >= "3.13"
pycparser==2.22 ; python_version >= "3.13"
pydantic-core==2.27.2 ; python_version >= "3.13"
pydantic==2.10.6 ; python_version >= "3.13"
pydub==0.25.1 ; python_version >= "3.13"
pymysql==1.1.1 ; python_version >= "3.13"
pyparsing==3.2.1 ; python_version >= "3.13"
python-dateutil==2.9.0.post0 ; python_version >= "3.13"
python-dotenv==1.0.1 ; python_version >= "3.13"
pytz==2025.1 ; python_version >= "3.13"
requests==2.32.3 ; python_version >= "3.13"
scikit-learn==1.6.1 ; python_version >= "3.13"
scipy==1.15.2 ; python_version >= "3.13"
seaborn==0.13.2 ; python_version >= "3.13"
six==1.17.0 ; python_version >= "3.13"
sniffio==1.3.1 ; python_version >= "3.13"
soundfile==0.13.1 ; python_version >= "3.13"
speechrecognition==3.14.1 ; python_version >= "3.13"
sqlalchemy==2.0.39 ; python_version >= "3.13"
standard-aifc==3.13.0 ; python_version >= "3.13"
standard-chunk==3.13.0 ; python_version >= "3.13"
statsmodels==0.14.4 ; python_version >= "3.13"
threadpoolctl==3.6.0 ; python_version >= "3.13"
tqdm==4.67.1 ; python_version >= "3.13"
typing-extensions==4.12.2 ; python_version >= "3.13"
tzdata==2025.1 ; python_version >= "3.13"
urllib3==2.3.0 ; python_version >= "3.13"
werkzeug==3.1.3 ; python_version >= "3.13"
